+++ tests/static_files/test_file_serving_resolution.py
@@ -7,9 +7,9 @@ from fsspec.implementations.local import LocalFileSystem
 
 from starlite import MediaType, get
 from starlite.config.static_files import StaticFilesConfig
+from starlite.file_system import BaseLocalFileSystem
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import create_test_client
-from starlite.utils.file import BaseLocalFileSystem
 
 if TYPE_CHECKING:
     from starlite.types import FileSystemProtocol
