+++ tests/test_response_containers.py
@@ -10,10 +10,10 @@ from fsspec.implementations.local import LocalFileSystem
 from starlite import get
 from starlite.datastructures import ETag
 from starlite.exceptions import ImproperlyConfiguredException
+from starlite.file_system import BaseLocalFileSystem
 from starlite.response_containers import File
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import RequestFactory, create_test_client
-from starlite.utils.file import BaseLocalFileSystem
 
 if TYPE_CHECKING:
     from starlite.types import FileSystemProtocol
